Widen scene-load notify context to uintptr for x64 (x64 fix)
The world/scene load-progress callback (PFNWORLDLOADNOTIFY) carried its
context as a uint32 dwContext, and the loader chain passed the context
(a CLoaderWnd* for the progress dialog) through uint32 at every hop:
tpassglobals (g_u4NotifyParam + the (uint32)pdlg casts + TPassLoadNotify)
-> CWorld::bLoadScene/bLoadWorld -> CLoadWorld -> CGroffFile / ImageLoader.
On x64 the dialog pointer was truncated to 32 bits, so TPassLoadNotify's
(CLoaderWnd*)dwContext gave a garbage pointer -> crash on interactive
loads (the autoload path passes context 0, which is why headless startup
survived). Widen only the context (dwContext / *NotifyParam / the
(uintptr)pdlg casts / g_u4NotifyParam and its Direct3D.cpp extern) to
uintptr; dwParam1/2/3 stay uint32 (they carry progress ints). Also fixes
the 8th/9th PFNWORLDLOADNOTIFY typedefs in Direct3D.cpp and GUIAppDlg.cpp
(the .cpp copies my header grep first missed). uintptr==uint32 on Win32
(build green); x64 clean + C4312-free.

+++ jp2_pc/Source/GUIApp/GUIAppDlg.cpp
@@ -209,7 +209,7 @@
 
 
 #ifndef PFNWORLDLOADNOTIFY
-typedef uint32 (__stdcall * PFNWORLDLOADNOTIFY)(uint32 dwContext, uint32 dwParam1, uint32 dwParam2, uint32 dwParam3);
+typedef uint32 (__stdcall * PFNWORLDLOADNOTIFY)(uintptr dwContext, uint32 dwParam1, uint32 dwParam2, uint32 dwParam3);
 #endif
 
 PFNWORLDLOADNOTIFY          g_pfnWorldLoadNotify;
